Reorganise c8y_mapper_ext::operations module
`operations` module was reorganised to better separate different
concerns:
- plain conversions of operations-related messages
- spawning and joining of operation tasks
- handling different types of operations
