+++ crates/extensions/c8y_mapper_ext/src/operations/handlers/software_update.rs
